Multiple changes
- cleaned up bounding box implementation by getting rid of unused
`_edit_mode` and `_keyboard_timer` attributes
- noticed that -- perhaps because of PySide2 migration? -- selecting
multiple bounding boxes with a mouse rubber band drag caused them to
lose keyboard focus. As such, when the “delete” key was pressed, the
bounding box would no longer execute its keyPressEvent and delete
itself. Same applies to moving bounding boxes with the keyboard keys.
For this reason, I moved all this functionality to the WGraphicsScene.
- bug fix: when saving a Tensorflow Record, do not save bounding boxes
for which their `delete` parameter is True